PR #88 introduced the use of an local HTTP server for JWKS key look up in unittests.  On windows this generates firewall warnings for each unittest which results in a popup box. This isn't a serious issue but it is annoying.

This PR switches the HTTP server for an httptest server which no longer triggers these warnings. 

* This change does not appear may also make the unittests run slightly faster but the difference is small enough not to matter. See measurement below.

* We are using testhttp for the githubOP test so it makes sense to standard our unittesting on httptest. httptest provides nice TLS functionality for tests as well.

Current
```
ok      github.com/openpubkey/openpubkey/cert   0.551s
ok      github.com/openpubkey/openpubkey/client 0.833s
ok      github.com/openpubkey/openpubkey/client/providers       0.303s
ok      github.com/openpubkey/openpubkey/cosigner       1.304s
ok      github.com/openpubkey/openpubkey/examples/mfa/mfacosigner       0.703s
ok      github.com/openpubkey/openpubkey/examples/ssh/opkssh    0.466s
ok      github.com/openpubkey/openpubkey/examples/ssh/sshcert   0.431s
ok      github.com/openpubkey/openpubkey/examples/x509/ca       0.418s
ok      github.com/openpubkey/openpubkey/gq     1.064s
ok      github.com/openpubkey/openpubkey/pktoken        0.815s
ok      github.com/openpubkey/openpubkey/pktoken/simplejws      0.377s
ok      github.com/openpubkey/openpubkey/verifier       0.920s
```

After this PR
```
ok      github.com/openpubkey/openpubkey/cert   0.378s
ok      github.com/openpubkey/openpubkey/client 0.866s
ok      github.com/openpubkey/openpubkey/client/providers       0.213s
ok      github.com/openpubkey/openpubkey/cosigner       0.942s
ok      github.com/openpubkey/openpubkey/examples/mfa/mfacosigner       0.865s
ok      github.com/openpubkey/openpubkey/examples/ssh/opkssh    0.526s
ok      github.com/openpubkey/openpubkey/examples/ssh/sshcert   0.409s
ok      github.com/openpubkey/openpubkey/examples/x509/ca       0.383s
ok      github.com/openpubkey/openpubkey/gq     1.058s
ok      github.com/openpubkey/openpubkey/pktoken        0.515s
ok      github.com/openpubkey/openpubkey/pktoken/simplejws      0.284s
ok      github.com/openpubkey/openpubkey/verifier       0.617s
```